Corrections to Problem Questions
Three corrected files.

+++ Imogen Files/chapter 8 try it out 1.py
@@ -1,3 +1,3 @@
-no = float(raw_input("Which multiplication table would you like?"))
+no = int(raw_input("Which multiplication table would you like?"))
 for i in range(1, 13):
     print no, "x", i, "=", no * i
